+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog
 =========
+## Unreleased
+### Added
+- ElementAccessor traits for element centric services
+
 ## 1.0.0-rc.7 - 2018-01-17
 ### Changed
 - Index actions are more flexible when creating the data provider